Refactor Combat
This cleans up the combat loop to be easier to understand. We take a turn,
then publish the END_OF_TURN, which causes enemies to take their turn.

I think it would be better to make it more symmetrical -- like have each
entity take a turn, but this works for now.